refactor: move Operator2 to core module
