+++ pennylane/__init__.py
@@ -33,7 +33,6 @@ from pennylane import kernels
 from pennylane import math
 from pennylane import operation
 from pennylane.core import operator
-from pennylane import operation2
 from pennylane import allocation
 from pennylane.allocation import allocate, deallocate
 from pennylane import decomposition
