+++ pennylane/core/operator/__init__.py
@@ -182,8 +182,18 @@ these objects are located in ``pennylane.ops.qubit.attributes``, not ``pennylane
 """
 
 from .base import Operator, Operation
+from .base2 import Operator2
 from .channel import Channel
 from .cv import CV, CVObservable, CVOperation
 from .state_prep import StatePrepBase
 
-__all__ = ["Operator", "Operation", "Channel", "CV", "CVObservable", "CVOperation", "StatePrepBase"]
+__all__ = [
+    "Operator",
+    "Operator2",
+    "Operation",
+    "Channel",
+    "CV",
+    "CVObservable",
+    "CVOperation",
+    "StatePrepBase",
+]
