+++ tests/tape/test_qscript.py
@@ -19,10 +19,10 @@ import numpy as np
 import pytest
 
 import pennylane as qp
+from pennylane.core.operator.base2 import Operator2
 from pennylane.exceptions import PennyLaneDeprecationWarning
 from pennylane.measurements import Shots, StateMP
 from pennylane.operation import _UNSET_BATCH_SIZE
-from pennylane.operation2 import Operator2
 from pennylane.queuing import AnnotatedQueue
 from pennylane.tape.qscript import QuantumScript, process_queue
